@@ -23,7 +23,7 @@ dependencies {
 kotlin {
     explicitApi()
     jvmToolchain {
-        (this as JavaToolchainSpec).languageVersion.set(JavaLanguageVersion.of("8"))
+        languageVersion.set(JavaLanguageVersion.of("8"))
     }
 }
 
@@ -40,8 +40,8 @@ tasks.withType<KotlinCompile>().configureEach {
     kotlinOptions.apply {
         jvmTarget = "1.8"
         freeCompilerArgs += listOf("-Xjsr305=strict")
-        apiVersion = "1.4"
-        languageVersion = "1.4"
+        apiVersion = "1.5"
+        languageVersion = "1.7"
     }
 }
 
